Fix compile errors on macOS

+++ src/lgfx/v1/platforms/opencv/Panel_OpenCV.cpp
@@ -173,10 +173,10 @@ namespace lgfx
 
   void Panel_OpenCV::setWindow(uint_fast16_t xs, uint_fast16_t ys, uint_fast16_t xe, uint_fast16_t ye)
   {
-    xs = std::max(0u, std::min<uint_fast16_t>(_width  - 1, xs));
-    xe = std::max(0u, std::min<uint_fast16_t>(_width  - 1, xe));
-    ys = std::max(0u, std::min<uint_fast16_t>(_height - 1, ys));
-    ye = std::max(0u, std::min<uint_fast16_t>(_height - 1, ye));
+    xs = std::max(static_cast<uint_fast16_t>(0u), std::min<uint_fast16_t>(_width  - 1, xs));
+    xe = std::max(static_cast<uint_fast16_t>(0u), std::min<uint_fast16_t>(_width  - 1, xe));
+    ys = std::max(static_cast<uint_fast16_t>(0u), std::min<uint_fast16_t>(_height - 1, ys));
+    ye = std::max(static_cast<uint_fast16_t>(0u), std::min<uint_fast16_t>(_height - 1, ye));
     _xpos = xs;
     _xs = xs;
     _xe = xe;
